Finished and tested implementation of TableManager for LunchViewController
