+++ Glancer/Main.storyboard
@@ -1,13 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14113" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="dMb-BZ-as2">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14490.70" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="dMb-BZ-as2">
     <device id="retina4_7" orientation="portrait">
         <adaptation id="fullscreen"/>
     </device>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
-        <capability name="Aspect ratio constraints" minToolsVersion="5.1"/>
-        <capability name="Constraints with non-1.0 multipliers" minToolsVersion="5.1"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14490.49"/>
         <capability name="Stack View standard spacing" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
@@ -139,41 +137,6 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="maD-5r-XlB" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="1049" y="-281"/>
-        </scene>
-        <!--Messages-->
-        <scene sceneID="Idh-qh-gsZ">
-            <objects>
-                <viewController storyboardIdentifier="Lunch" title="Messages" id="nbf-qm-dY2" customClass="LunchController" customModule="Glancer" customModuleProvider="target" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="qjs-QZ-n1y"/>
-                        <viewControllerLayoutGuide type="bottom" id="alR-gk-ykM"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="V4c-Q9-BRu">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" showsVerticalScrollIndicator="NO" dataMode="prototypes" style="plain" separatorStyle="none" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="RhM-mi-FP3">
-                                <rect key="frame" x="0.0" y="64" width="375" height="603"/>
-                                <color key="backgroundColor" cocoaTouchSystemColor="groupTableViewBackgroundColor"/>
-                            </tableView>
-                        </subviews>
-                        <color key="backgroundColor" cocoaTouchSystemColor="groupTableViewBackgroundColor"/>
-                        <constraints>
-                            <constraint firstItem="RhM-mi-FP3" firstAttribute="leading" secondItem="V4c-Q9-BRu" secondAttribute="leading" id="2C7-BX-ySo"/>
-                            <constraint firstItem="RhM-mi-FP3" firstAttribute="top" secondItem="qjs-QZ-n1y" secondAttribute="bottom" id="7dm-Rl-fCN"/>
-                            <constraint firstItem="alR-gk-ykM" firstAttribute="top" secondItem="RhM-mi-FP3" secondAttribute="bottom" id="ABF-Qv-Gje"/>
-                            <constraint firstAttribute="trailing" secondItem="RhM-mi-FP3" secondAttribute="trailing" id="GCx-V5-SAt"/>
-                        </constraints>
-                    </view>
-                    <navigationItem key="navigationItem" title="Lunch" id="N9C-YP-6vD"/>
-                    <simulatedNavigationBarMetrics key="simulatedTopBarMetrics" prompted="NO"/>
-                    <connections>
-                        <outlet property="tableView" destination="RhM-mi-FP3" id="T5i-jg-pKE"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="J50-1Y-Fhb" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="292" y="-281"/>
         </scene>
         <!--Upcoming Navigation-->
         <scene sceneID="Rr5-Ct-Wsk">
